Trait extract_map::LendingIterator

pub trait LendingIterator {
    type Item<'a>
       where Self: 'a;

    // Required method
    fn next(&mut self) -> Option<Self::Item<'_>>;

    // Provided methods
    fn size_hint(&self) -> (usize, Option<usize>) { ... }
    fn count(self) -> usize
       where Self: Sized { ... }
    fn advance_by(&mut self, n: usize) -> Result<(), NonZero<usize>> { ... }
    fn nth(&mut self, n: usize) -> Option<Self::Item<'_>> { ... }
    fn step_by(self, step: usize) -> StepBy<Self>
       where Self: Sized { ... }
    fn take(self, n: usize) -> Take<Self>
       where Self: Sized { ... }
    fn take_while<P>(self, predicate: P) -> TakeWhile<Self, P>
       where Self: Sized,
             P: for<'a> FnMut(&Self::Item<'a>) -> bool { ... }
    fn chain<I>(self, other: I) -> Chain<Self, I>
       where Self: Sized,
             I: for<'a> LendingIterator<Item<'a> = Self::Item<'a>> + 'a { ... }
    fn zip<I>(self, other: I) -> Zip<Self, I>
       where Self: Sized,
             I: LendingIterator { ... }
    fn map<F>(self, f: F) -> Map<Self, F>
       where Self: Sized,
             F: for<'a> SingleArgFnMut<Self::Item<'a>> { ... }
    fn for_each<F>(self, f: F)
       where Self: Sized,
             F: FnMut(Self::Item<'_>) { ... }
    fn filter<P>(self, predicate: P) -> Filter<Self, P>
       where Self: Sized,
             P: for<'a> FnMut(&Self::Item<'a>) -> bool { ... }
    fn filter_map<F>(self, f: F) -> FilterMap<Self, F>
       where Self: Sized,
             F: for<'a> SingleArgFnMut<Self::Item<'a>>,
             <F as SingleArgFnOnce<Self::Item<'a>>>::Output: for<'a> OptionTrait { ... }
    fn fold<B, F>(self, init: B, f: F) -> B
       where Self: Sized,
             F: FnMut(B, Self::Item<'_>) -> B { ... }
    fn cloned<T>(self) -> Cloned<Self>
       where Self: Sized,
             Self::Item<'a>: for<'a> Deref<Target = T>,
             T: Clone { ... }
    fn enumerate(self) -> Enumerate<Self>
       where Self: Sized { ... }
    fn skip(self, n: usize) -> Skip<Self>
       where Self: Sized { ... }

}

Like Iterator, but items may borrow from &mut self.

This means that the compiler will check that you finish using an item before requesting the next item, as it’s not allowed for two &mut self to exist at the same time.

Required Associated Types

type Item<'a> where Self: 'a

The type of the elements being iterated over.

Required Methods

fn next(&mut self) -> Option<Self::Item<'_>>

Advances the lending iterator and returns the next value.

See Iterator::next.

Provided Methods

fn size_hint(&self) -> (usize, Option<usize>)

Returns the bounds on the remaining length of the iterator.

See Iterator::size_hint.

fn count(self) -> usize
where Self: Sized,

Returns the number of items in the lending iterator.

See Iterator::count.

fn advance_by(&mut self, n: usize) -> Result<(), NonZero<usize>>

Advances the lending iterator by n elements.

See Iterator::advance_by.

fn nth(&mut self, n: usize) -> Option<Self::Item<'_>>

Returns the nth element of the lending iterator.

See Iterator::nth.

fn step_by(self, step: usize) -> StepBy<Self>
where Self: Sized,

Creates a lending iterator starting at the same point, but stepping by the given amount at each iteration.

See Iterator::step_by.

fn take(self, n: usize) -> Take<Self>
where Self: Sized,

Creates a lending iterator that lends the first n elements, or fewer if the underlying iterator ends sooner.

See Iterator::take.

fn take_while<P>(self, predicate: P) -> TakeWhile<Self, P>
where Self: Sized, P: for<'a> FnMut(&Self::Item<'a>) -> bool,

Creates a lending iterator that lends items matching a predicate.

The predicate is called once for every item. Once it returns false once, None is returned for all subsequent calls to next.

fn chain<I>(self, other: I) -> Chain<Self, I>
where Self: Sized, I: for<'a> LendingIterator<Item<'a> = Self::Item<'a>> + 'a,

Takes two lending iterators and creates a new lending iterator over both in sequence.

See Iterator::chain.

fn zip<I>(self, other: I) -> Zip<Self, I>
where Self: Sized, I: LendingIterator,

‘Zips up’ two lending iterators into a single lending iterator of pairs.

fn map<F>(self, f: F) -> Map<Self, F>
where Self: Sized, F: for<'a> SingleArgFnMut<Self::Item<'a>>,

Takes a closure and creates a lending iterator which calls that closure on each element.

As of writing, in stable rust it’s not possible to create a closure where the lifetime of its output is tied to its input. If you’re on nightly, you can use the unstable closure_lifetime_binder feature. If you’re on stable, try using a function.

In the case that the closure’s return type doesn’t borrow from its input, the resulting LendingIterator will implement IntoIterator.

See Iterator::map.

fn for_each<F>(self, f: F)
where Self: Sized, F: FnMut(Self::Item<'_>),

Calls a closure on each element of the lending iterator.

See Iterator::for_each.

fn filter<P>(self, predicate: P) -> Filter<Self, P>
where Self: Sized, P: for<'a> FnMut(&Self::Item<'a>) -> bool,

Creates a lending iterator which uses a closure to determine if an element should be yielded.

See Iterator::filter.

fn filter_map<F>(self, f: F) -> FilterMap<Self, F>
where Self: Sized, F: for<'a> SingleArgFnMut<Self::Item<'a>>, <F as SingleArgFnOnce<Self::Item<'a>>>::Output: for<'a> OptionTrait,

Creates a lending iterator that both filters and maps.

See Iterator::filter_map.

fn fold<B, F>(self, init: B, f: F) -> B
where Self: Sized, F: FnMut(B, Self::Item<'_>) -> B,

Folds every element into an accumulator by applying an operation, returning the final result.

See Iterator::fold.

fn cloned<T>(self) -> Cloned<Self>
where Self: Sized, Self::Item<'a>: for<'a> Deref<Target = T>, T: Clone,

Creates a lending iterator which clones all of its elements.

The resulting lending iterator implements IntoIterator.

See Iterator::cloned.

fn enumerate(self) -> Enumerate<Self>
where Self: Sized,

Creates a lending iterator which gives the current iteration count as well as the next value.

fn skip(self, n: usize) -> Skip<Self>
where Self: Sized,

Creates a lending iterator that skips over the first n elements of self.

Object Safety

This trait is not object safe.